merging develop to main for deployment

+++ controllers/cocktailcontroller.js
@@ -1,12 +1,12 @@
-const express = require('express')
-const router = express.Router()
-const Cocktail = require('../db').import('../models/cocktail')
+const express = require('express');
+const router = express.Router();
+const validateSession = require('../middleware/validate-session');
+const Cocktail = require('../db').import('../models/cocktail');
 
 /************************
     * RECIPE CREATE *
 *************************/
-//change route and user id req once ready for validation
-router.post('/:id', (req, res) => {
+router.post('/',validateSession, ( req, res) => {
     const cocktailEntry = {
         name: req.body.name,
         alcoholic: req.body.alcoholic,
@@ -17,8 +17,7 @@ router.post('/:id', (req, res) => {
         iced: req.body.iced,
         shaken: req.body.shaken,
         stirred: req.body.stirred,
-        //userId: req.user.id - to use once we have validate session
-        userId: req.params.id
+        userId: req.user.id,
     }
     Cocktail.create(cocktailEntry)
         .then(logs => res.status(200).json(logs))
@@ -28,8 +27,8 @@ router.post('/:id', (req, res) => {
 /*******************************
   * GET ALL COCKTAILS BY USER *
 ********************************/
-router.get("/:id", (req, res) => {
-    Cocktail.findAll({ where: { userId: req.params.id } })
+router.get("/", validateSession, (req, res) => {
+    Cocktail.findAll({ where: { userId: req.user.id } })
         .then((cocktails) => {
             if (cocktails.length === 0)
                 return res.status(200).json({ message: "No cocktails were found! Try creating one." });
@@ -58,9 +57,20 @@ router.get("/cocktails/:name", (req, res) => {
 /*******************************
    * UPDATE COCKTAIL RECIPE *
 ********************************/
-router.put("/:id", (req, res) => {
-    Cocktail.update(req.body, {
-        where: { id: req.params.id },
+router.put("/:id", validateSession, (req, res) => {
+    const cocktailUpdate = {
+        name: req.body.name,
+        alcoholic: req.body.alcoholic,
+        glassType: req.body.glassType,
+        ingredients: req.body.ingredients,
+        measurements: req.body.measurements,
+        instructions: req.body.instructions,
+        iced: req.body.iced,
+        shaken: req.body.shaken,
+        stirred: req.body.stirred
+    }
+    Cocktail.update(cocktailUpdate, {
+        where: { id: req.params.id, userId: req.user.id },
     })
         .then((response) => {
             res.status(200).json({ message: "Your cocktail has been updated.", response });
@@ -73,9 +83,8 @@ router.put("/:id", (req, res) => {
 /*************************
     * DELETE COCKTAIL *
 **************************/
-//add second where 'userId: req.user.id' once we add validate session
-router.delete("/:id", (req, res) => {
-    Cocktail.destroy({ where: { id: req.params.id } })
+router.delete("/:id", validateSession, (req, res) => {
+    Cocktail.destroy({ where: { id: req.params.id, userId: req.user.id } })
     .then((result) => {
         if (result) {
             return res.status(200).json({ message: `Successfully deleted ${result}`})
